Removed install.py and installer.sh and moved into helper scripts
Was running into problems on some systems with the installed script not
being able to find the virtual environment site packages when using the
removed installers. It works for some reason when the installers are
sourced into the current shell session, so moved to there until I can
figure out what's going on. Also deleted some code, so bonus.